requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
multidict==6.5.0
pillow==11.2.1
propcache==0.3.2
pycparser==2.22
pyperclip==1.9.0
python-dotenv==1.1.0
requests==2.32.4
urllib3==2.5.0
yarl==1.20.1
